+++ CHANGELOG.rst
@@ -16,6 +16,8 @@ Changelog
 * Update the :command:`fuc-demux` command to output a better figure.
 * Add new method :meth:`common.plot_cytobands`.
 * Add new method :meth:`pycov.CovFrame.plot_uniformity`.
+* Rename :meth:`pycov.CovFrame.from_file` method to :meth:`pycov.CovFrame.from_bam`.
+* Add new command :command:`fuc-depth`.
 
 0.18.0 (2021-07-20)
 -------------------